requirements.txt
asgiref==3.4.1
astroid==2.8.5
black==21.10b0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Django==3.2.9
django-environ==0.8.1
django-isbn-field==0.5.2
django-language-field==0.0.3
djangorestframework==3.12.4
flake8==4.0.1
idna==3.3
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
python-stdnum==1.17
pytz==2021.3
regex==2021.11.10
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
typing-extensions==4.0.0
urllib3==1.26.7
wrapt==1.13.3
